refactor(rag): extract LLM rerank into rag/rerank.rs
Move parse_rerank_indices and llm_rerank into rerank.rs. Drop now-unused
RERANK_PROMPT import from mod.rs. Behavior unchanged.
